coder/pyproject.toml
[project]
name = "agentic-python-coder"
version = "0.2.0"
description = "A lightweight Python coding agent that writes, executes, and iterates on code through natural language instructions"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.6.6",
    "langchain-core>=0.3.74",
    "langchain-openai>=0.3.31",
    "langchain-anthropic>=0.3.19",
    "langchain-experimental>=0.3.4",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "jupyter-client>=8.6.3",
    "ipykernel>=6.30.1",
    "PyYAML>=6.0.2",
]

[project.scripts]
coder = "coder.cli:main"

[project.optional-dependencies]
test = [
    "pytest>=8.4.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/coder"]

[tool.pytest.ini_options]
pythonpath = ["src"]

